Final commit

+++ SpringBoot/hw5/shopping-cart/shopping-cart/src/main/java/com/store/shopping_cart/controller/ProductController.java
@@ -2,6 +2,8 @@ package com.store.shopping_cart.controller;
 
 import java.util.List;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -9,6 +11,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.store.shopping_cart.model.Product;
@@ -29,15 +32,25 @@ public class ProductController {
         return productService.getProducts();
     }
 
-    @PutMapping(value = "/products/{id}")
-    public Product updateProduct(@PathVariable Integer id, @RequestBody Product product) {
-        return productService.updateProduct(id, product);
+    @PutMapping(value = "/products/update")
+    public ResponseEntity<Object> updateProduct(@RequestParam("id") Integer id, @RequestBody Product product) {
+        if (productService.productExists(id)) {
+            productService.updateProduct(id, product);
+            return new ResponseEntity<>(product, HttpStatus.OK);
+        }
+        
+        return new ResponseEntity<>("Product does not exist", HttpStatus.NOT_FOUND);
      
     }
 
     @DeleteMapping(value = "/products/{id}")
-    public void deleteProduct(@PathVariable Integer id) {
-        productService.deleteProduct(id);
+    public ResponseEntity<Object> deleteProduct(@PathVariable Integer id) {
+         if (productService.productExists(id)) {
+          productService.deleteProduct(id);
+            return new ResponseEntity<>("Product " + id + " has been deleted.", HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>("Product does not exist", HttpStatus.NOT_FOUND);
 
     }
 
@@ -47,9 +60,17 @@ public class ProductController {
       return productService.createProduct(product);
     }
 
-    @GetMapping(value = "/products/id/{id}")
-    public Product findProductByID(@PathVariable int id) {
-        return productService.getProduct(id);
+    @GetMapping(value = "/products/find")
+    public ResponseEntity<Object> findProductByID(@RequestParam("id") int id) {
+
+         if (productService.productExists(id)) {
+          Product product = productService.getProduct(id);
+            return new ResponseEntity<>(product, HttpStatus.OK);
+        }
+        
+        return new ResponseEntity<>("Product does not exist", HttpStatus.NOT_FOUND);
+
+    
     }
 
 }
